Ignore coverage folder to avoide senseless changes
We probably don't want to check in changes to the index.html,
resultset.json etc. to this repo. It's also not recommended
to check in by the maintainers. There are arguments for
checking in `.last_run.json` but I personally wouldn't recommend
it as updating it is also easy to forget.

It's also annoying to have all these changes here after running
the tests.