Adjust the order in which components are provided
Before it was rather confusing for me as other component doesn't
use some component and later also isn't used in the page so it
was rather difficult to find out how both components end up
in the page.

I hope that by reversing the order of the components so that first
people are introduced to other_component, which is then used in
some component which is then used in the page this is made more
understandable.

Also added another note mentioning this.