Remove unused render code + add security warning for isolate
The old code was just still there as an orientation help for
rendering, not needed anymore. New code is/should be much
more intuitive and already points out where the remaining things
should go.

Also added a security notice that occurred to me while thinking
about the isolate component and ways in which it mgiht break/
be a security vulnerability as letting the front end tell you
which (Ruby) class to render is an inherent security risk
and we might need to come up with safeguards/other concepts
around that.